Legendary aspect tracking, and spell management for active creatures.
Set up radios for tracking legendary resistances and legendary actions.
Made legendary actions reset at the start of its turn.
Set up dropdowns for spells creatures have access to, includes radios
for tracking individual creatures slots/uses.

Set up quick has methods to see if a spell list has daily or prepared
spells.

+++ Encounters/EncounterBuilder.java
@@ -659,7 +659,7 @@ public class EncounterBuilder {
 						@Override
 						public void handle(ActionEvent event) {
 								double[] custom = multipliers.get("Custom multiplier");
-								for(int i=0; i<custom.length; i++) {custom[i] = Integer.parseInt(customMults[i].getText());}
+								for(int i=0; i<custom.length; i++) {custom[i] = Double.parseDouble(customMults[i].getText());}
 								updateXPPanel();
 							
 						}
